requirements.txt
numpy==1.21.5
pygame==2.2.0
keras==2.9.0
tensorflow==2.9.1
neat-python==0.92